+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This project follows semantic versioning.
 
 ### Unpublished
+    - [added] Type-level arrays of type-level integers. (PR #66)
     - [added] The types in this crate are now instantiable. (Issue #67, PR #68)
 
 ### 1.3.1 (2016-03-31)